now displays pence

+++ db/book_shop.sql
@@ -19,8 +19,8 @@ CREATE TABLE books (
     author VARCHAR(255),
     publisher_id INT REFERENCES publishers(id),
     genre_id INT REFERENCES genres(id),
-    buying_cost INT,
-    price INT,
+    buying_cost FLOAT,
+    price FLOAT,
     stock INT,
     description VARCHAR(2000)
